Factoring out the cache code into it's own file
