+++ lib/services_gae.dart
@@ -8,7 +8,6 @@ import 'dart:async';
 import 'dart:io' as io;
 
 import 'package:appengine/appengine.dart' as ae;
-import 'package:dart_services/src/sdk_manager.dart';
 import 'package:logging/logging.dart';
 import 'package:rpc/rpc.dart' as rpc;
 
@@ -16,12 +15,14 @@ import 'src/common.dart';
 import 'src/common_server.dart';
 import 'src/dartpad_support_server.dart';
 import 'src/flutter_web.dart';
+import 'src/server_cache.dart';
+import 'src/sdk_manager.dart';
 
 const String _API = '/api';
 const String _healthCheck = '/_ah/health';
 const String _readynessCheck = '/_ah/ready';
 
-final Logger _logger = Logger('gae_server');
+final Logger _log = Logger('gae_server');
 
 void main(List<String> args) {
   var gaePort = 8080;
@@ -35,7 +36,7 @@ void main(List<String> args) {
 
   // Log to stdout/stderr.  AppEngine's logging package is disabled in 0.6.0
   // and AppEngine copies stdout/stderr to the dashboards.
-  _logger.onRecord.listen((LogRecord rec) {
+  _log.onRecord.listen((LogRecord rec) {
     final out = ('${rec.level.name}: ${rec.time}: ${rec.message}\n');
     if (rec.level > Level.INFO) {
       io.stderr.write(out);
@@ -43,7 +44,7 @@ void main(List<String> args) {
       io.stdout.write(out);
     }
   });
-  log.info('''Initializing dart-services:
+  _log.info('''Initializing dart-services:
     port: $gaePort
     sdkPath: $sdkPath
     REDIS_SERVER_URI: ${io.Platform.environment['REDIS_SERVER_URI']}
@@ -66,7 +67,7 @@ class GaeServer {
 
   GaeServer(this.sdkPath, this.redisServerUri) {
     hierarchicalLoggingEnabled = true;
-    _logger.level = Level.ALL;
+    _log.level = Level.ALL;
 
     discoveryEnabled = false;
     fileRelayServer = FileRelayServer();
@@ -129,7 +130,7 @@ class GaeServer {
 
   Future _processHealthRequest(io.HttpRequest request) async {
     if (commonServer.running && !commonServer.analysisServersRunning) {
-      _logger.severe('CommonServer running without analysis servers. '
+      _log.severe('CommonServer running without analysis servers. '
           'Intentionally failing healthcheck.');
       request.response.statusCode = io.HttpStatus.internalServerError;
     } else {
@@ -148,7 +149,7 @@ class GaeServer {
           await tempDir.delete(recursive: true);
         }
       } catch (e) {
-        _logger.severe('Failed to create temporary file: $e');
+        _log.severe('Failed to create temporary file: $e');
         request.response.statusCode = io.HttpStatus.internalServerError;
       }
     }
@@ -175,7 +176,7 @@ class GaeServer {
     } catch (e) {
       // This should only happen in the case where there is a bug in the rpc
       // package. Otherwise it always returns an HttpApiResponse.
-      _logger.warning('Failed with error: $e when trying to call '
+      _log.warning('Failed with error: $e when trying to call '
           'method at \'${request.uri.path}\'.');
       request.response.statusCode = io.HttpStatus.internalServerError;
       await request.response.close();
