+++ lib/src/common_server.dart
@@ -7,13 +7,10 @@ library services.common_server;
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
-import 'dart:math';
 
 import 'package:crypto/crypto.dart';
-import 'package:dartis/dartis.dart' as redis;
 import 'package:logging/logging.dart';
 import 'package:pedantic/pedantic.dart';
-import 'package:quiver/cache.dart';
 import 'package:rpc/rpc.dart';
 
 import '../version.dart';
@@ -23,20 +20,11 @@ import 'common.dart';
 import 'compiler.dart';
 import 'flutter_web.dart';
 import 'pub.dart';
+import 'server_cache.dart';
 import 'sdk_manager.dart';
 
 final Duration _standardExpiration = Duration(hours: 1);
-final Logger log = Logger('common_server');
-
-abstract class ServerCache {
-  Future<String> get(String key);
-
-  Future<void> set(String key, String value, {Duration expiration});
-
-  Future<void> remove(String key);
-
-  Future<void> shutdown();
-}
+final Logger _log = Logger('common_server');
 
 abstract class ServerContainer {
   String get version;
@@ -46,215 +34,6 @@ class SummaryText {
   String text;
 
   SummaryText.fromString(this.text);
-}
-
-/// A redis-backed implementation of [ServerCache].
-class RedisCache implements ServerCache {
-  redis.Client redisClient;
-  redis.Connection _connection;
-
-  final String redisUriString;
-
-  // Version of the server to add with keys.
-  final String serverVersion;
-
-  // pseudo-random is good enough.
-  final Random randomSource = Random();
-  static const int _connectionRetryBaseMs = 250;
-  static const int _connectionRetryMaxMs = 60000;
-  static const Duration cacheOperationTimeout = Duration(milliseconds: 10000);
-
-  RedisCache(this.redisUriString, this.serverVersion) {
-    _reconnect();
-  }
-
-  Completer<void> _connected = Completer<void>();
-
-  /// Completes when and if the redis server connects.  This future is reset
-  /// on disconnection.  Mostly for testing.
-  Future<void> get connected => _connected.future;
-
-  Completer<void> _disconnected = Completer<void>()..complete();
-
-  /// Completes when the server is disconnected (begins completed).  This
-  /// future is reset on connection.  Mostly for testing.
-  Future<void> get disconnected => _disconnected.future;
-
-  String __logPrefix;
-
-  String get _logPrefix =>
-      __logPrefix ??= 'RedisCache [$redisUriString] ($serverVersion)';
-
-  bool _isConnected() => redisClient != null && !_isShutdown;
-  bool _isShutdown = false;
-
-  /// If you will no longer be using the [RedisCache] instance, call this to
-  /// prevent reconnection attempts.  All calls to get/remove/set on this object
-  /// will return null after this.  Future completes when disconnection is complete.
-  @override
-  Future<void> shutdown() {
-    log.info('$_logPrefix: shutting down...');
-    _isShutdown = true;
-    redisClient?.disconnect();
-    return disconnected;
-  }
-
-  /// Call when an active connection has disconnected.
-  void _resetConnection() {
-    assert(_connected.isCompleted && !_disconnected.isCompleted);
-    _connected = Completer<void>();
-    _connection = null;
-    redisClient = null;
-    _disconnected.complete();
-  }
-
-  /// Call when a new connection is established.
-  void _setUpConnection(redis.Connection newConnection) {
-    assert(_disconnected.isCompleted && !_connected.isCompleted);
-    _disconnected = Completer<void>();
-    _connection = newConnection;
-    redisClient = redis.Client(_connection);
-    _connected.complete();
-  }
-
-  /// Begin a reconnection loop asynchronously to maintain a connection to the
-  /// redis server.  Never stops trying until shutdown() is called.
-  void _reconnect([int retryTimeoutMs = _connectionRetryBaseMs]) {
-    if (_isShutdown) {
-      return;
-    }
-    log.info('$_logPrefix: reconnecting to $redisUriString...');
-    var nextRetryMs = retryTimeoutMs;
-    if (retryTimeoutMs < _connectionRetryMaxMs / 2) {
-      // 1 <= (randomSource.nextDouble() + 1) < 2
-      nextRetryMs = (retryTimeoutMs * (randomSource.nextDouble() + 1)).toInt();
-    }
-    redis.Connection.connect(redisUriString)
-        .then((redis.Connection newConnection) {
-          log.info('$_logPrefix: Connected to redis server');
-          _setUpConnection(newConnection);
-          // If the client disconnects, discard the client and try to connect again.
-          newConnection.done.then((_) {
-            _resetConnection();
-            log.warning('$_logPrefix: connection terminated, reconnecting');
-            _reconnect();
-          }).catchError((dynamic e) {
-            _resetConnection();
-            log.warning(
-                '$_logPrefix: connection terminated with error $e, reconnecting');
-            _reconnect();
-          });
-        })
-        .timeout(Duration(milliseconds: _connectionRetryMaxMs))
-        .catchError((_) {
-          log.severe(
-              '$_logPrefix: Unable to connect to redis server, reconnecting in ${nextRetryMs}ms ...');
-          Future<void>.delayed(Duration(milliseconds: nextRetryMs)).then((_) {
-            _reconnect(nextRetryMs);
-          });
-        });
-  }
-
-  /// Build a key that includes the server version, Dart SDK version, and
-  /// Flutter SDK version.
-  ///
-  /// We don't use the existing key directly so that different AppEngine
-  /// versions using the same redis cache do not have collisions.
-  String _genKey(String key) =>
-      'server:$serverVersion:dart:${SdkManager.sdk.versionFull}:flutter:${SdkManager.flutterSdk.versionFull}+$key';
-
-  @override
-  Future<String> get(String key) async {
-    String value;
-    key = _genKey(key);
-    if (!_isConnected()) {
-      log.warning('$_logPrefix: no cache available when getting key $key');
-    } else {
-      final commands = redisClient.asCommands<String, String>();
-      // commands can return errors synchronously in timeout cases.
-      try {
-        value = await commands.get(key).timeout(cacheOperationTimeout,
-            onTimeout: () async {
-          log.warning('$_logPrefix: timeout on get operation for key $key');
-          await redisClient?.disconnect();
-          return null;
-        });
-      } catch (e) {
-        log.warning('$_logPrefix: error on get operation for key $key: $e');
-      }
-    }
-    return value;
-  }
-
-  @override
-  Future<dynamic> remove(String key) async {
-    key = _genKey(key);
-    if (!_isConnected()) {
-      log.warning('$_logPrefix: no cache available when removing key $key');
-      return null;
-    }
-
-    final commands = redisClient.asCommands<String, String>();
-    // commands can sometimes return errors synchronously in timeout cases.
-    try {
-      return commands.del(key: key).timeout(cacheOperationTimeout,
-          onTimeout: () async {
-        log.warning('$_logPrefix: timeout on remove operation for key $key');
-        await redisClient?.disconnect();
-        return null;
-      });
-    } catch (e) {
-      log.warning('$_logPrefix: error on remove operation for key $key: $e');
-    }
-  }
-
-  @override
-  Future<void> set(String key, String value, {Duration expiration}) async {
-    key = _genKey(key);
-    if (!_isConnected()) {
-      log.warning('$_logPrefix: no cache available when setting key $key');
-      return null;
-    }
-
-    final commands = redisClient.asCommands<String, String>();
-    // commands can sometimes return errors synchronously in timeout cases.
-    try {
-      return Future<void>.sync(() async {
-        await commands.multi();
-        unawaited(commands.set(key, value));
-        if (expiration != null) {
-          unawaited(commands.pexpire(key, expiration.inMilliseconds));
-        }
-        await commands.exec();
-      }).timeout(cacheOperationTimeout, onTimeout: () {
-        log.warning('$_logPrefix: timeout on set operation for key $key');
-        redisClient?.disconnect();
-      });
-    } catch (e) {
-      log.warning('$_logPrefix: error on set operation for key $key: $e');
-    }
-  }
-}
-
-/// An in-memory implementation of [ServerCache] which doesn't support
-/// expiration of entries based on time.
-class InMemoryCache implements ServerCache {
-  /// Wrapping an internal cache with a maximum size of 512 entries.
-  final Cache<String, String> _lru =
-      MapCache<String, String>.lru(maximumSize: 512);
-
-  @override
-  Future<String> get(String key) async => _lru.get(key);
-
-  @override
-  Future<void> set(String key, String value, {Duration expiration}) async =>
-      _lru.set(key, value);
-
-  @override
-  Future<void> remove(String key) async => _lru.invalidate(key);
-
-  @override
-  Future<void> shutdown() => Future<void>.value();
 }
 
 @ApiClass(name: 'dartservices', version: 'v1')
@@ -281,11 +60,11 @@ class CommonServer {
     this.cache,
   ) {
     hierarchicalLoggingEnabled = true;
-    log.level = Level.ALL;
+    _log.level = Level.ALL;
   }
 
   Future<void> init() async {
-    log.info('Beginning CommonServer init().');
+    _log.info('Beginning CommonServer init().');
     analysisServer = AnalysisServerWrapper(sdkPath, flutterWebManager);
     flutterAnalysisServer = AnalysisServerWrapper(
         flutterWebManager.flutterSdk.sdkPath, flutterWebManager);
@@ -294,20 +73,20 @@ class CommonServer {
         Compiler(SdkManager.sdk, SdkManager.flutterSdk, flutterWebManager);
 
     await analysisServer.init();
-    log.info('Dart analysis server initialized.');
+    _log.info('Dart analysis server initialized.');
 
     await flutterAnalysisServer.init();
-    log.info('Flutter analysis server initialized.');
+    _log.info('Flutter analysis server initialized.');
 
     unawaited(analysisServer.onExit.then((int code) {
-      log.severe('analysisServer exited, code: $code');
+      _log.severe('analysisServer exited, code: $code');
       if (code != 0) {
         exit(code);
       }
     }));
 
     unawaited(flutterAnalysisServer.onExit.then((int code) {
-      log.severe('flutterAnalysisServer exited, code: $code');
+      _log.severe('flutterAnalysisServer exited, code: $code');
       if (code != 0) {
         exit(code);
       }
@@ -323,14 +102,14 @@ class CommonServer {
   }
 
   Future<void> restart() async {
-    log.warning('Restarting CommonServer');
+    _log.warning('Restarting CommonServer');
     await shutdown();
-    log.info('Analysis Servers shutdown');
+    _log.info('Analysis Servers shutdown');
 
     await init();
     await warmup();
 
-    log.warning('Restart complete');
+    _log.warning('Restart complete');
   }
 
   Future<dynamic> shutdown() {
@@ -448,10 +227,10 @@ class CommonServer {
       final results = await getCorrectAnalysisServer(source).analyze(source);
       final lineCount = source.split('\n').length;
       final ms = watch.elapsedMilliseconds;
-      log.info('PERF: Analyzed $lineCount lines of Dart in ${ms}ms.');
+      _log.info('PERF: Analyzed $lineCount lines of Dart in ${ms}ms.');
       return results;
     } catch (e, st) {
-      log.severe('Error during analyze', e, st);
+      _log.severe('Error during analyze', e, st);
       await restart();
       rethrow;
     }
@@ -473,7 +252,7 @@ class CommonServer {
 
     final result = await checkCache(memCacheKey);
     if (result != null) {
-      log.info('CACHE: Cache hit for compileDart2js');
+      _log.info('CACHE: Cache hit for compileDart2js');
       final resultObj = JsonDecoder().convert(result);
       return CompileResponse(
         resultObj['compiledJS'] as String,
@@ -481,7 +260,7 @@ class CommonServer {
       );
     }
 
-    log.info('CACHE: MISS for compileDart2js');
+    _log.info('CACHE: MISS for compileDart2js');
     final watch = Stopwatch()..start();
 
     return compiler
@@ -491,7 +270,7 @@ class CommonServer {
         final lineCount = source.split('\n').length;
         final outputSize = (results.compiledJS.length / 1024).ceil();
         final ms = watch.elapsedMilliseconds;
-        log.info('PERF: Compiled $lineCount lines of Dart into '
+        _log.info('PERF: Compiled $lineCount lines of Dart into '
             '${outputSize}kb of JavaScript in ${ms}ms using dart2js.');
         final sourceMap = returnSourceMap ? results.sourceMap : null;
 
@@ -509,7 +288,7 @@ class CommonServer {
       }
     }).catchError((dynamic e, dynamic st) {
       if (e is! BadRequestError) {
-        log.severe('Error during compile (dart2js): $e\n$st');
+        _log.severe('Error during compile (dart2js): $e\n$st');
       }
       throw e;
     });
@@ -527,7 +306,7 @@ class CommonServer {
 
     final result = await checkCache(memCacheKey);
     if (result != null) {
-      log.info('CACHE: Cache hit for compileDDC');
+      _log.info('CACHE: Cache hit for compileDDC');
       final resultObj = JsonDecoder().convert(result);
       return CompileDDCResponse(
         resultObj['compiledJS'] as String,
@@ -535,7 +314,7 @@ class CommonServer {
       );
     }
 
-    log.info('CACHE: MISS for compileDDC');
+    _log.info('CACHE: MISS for compileDDC');
     final watch = Stopwatch()..start();
 
     return compiler.compileDDC(source).then((DDCCompilationResults results) {
@@ -543,7 +322,7 @@ class CommonServer {
         final lineCount = source.split('\n').length;
         final outputSize = (results.compiledJS.length / 1024).ceil();
         final ms = watch.elapsedMilliseconds;
-        log.info('PERF: Compiled $lineCount lines of Dart into '
+        _log.info('PERF: Compiled $lineCount lines of Dart into '
             '${outputSize}kb of JavaScript in ${ms}ms using DDC.');
 
         final cachedResult = JsonEncoder().convert(<String, String>{
@@ -560,7 +339,7 @@ class CommonServer {
       }
     }).catchError((dynamic e, dynamic st) {
       if (e is! BadRequestError) {
-        log.severe('Error during compile (DDC): $e\n$st');
+        _log.severe('Error during compile (DDC): $e\n$st');
       }
       throw e;
     });
@@ -581,10 +360,10 @@ class CommonServer {
       var docInfo =
           await getCorrectAnalysisServer(source).dartdoc(source, offset);
       docInfo ??= <String, String>{};
-      log.info('PERF: Computed dartdoc in ${watch.elapsedMilliseconds}ms.');
+      _log.info('PERF: Computed dartdoc in ${watch.elapsedMilliseconds}ms.');
       return DocumentResponse(docInfo);
     } catch (e, st) {
-      log.severe('Error during dartdoc', e, st);
+      _log.severe('Error during dartdoc', e, st);
       await restart();
       rethrow;
     }
@@ -611,10 +390,10 @@ class CommonServer {
     try {
       final response =
           await getCorrectAnalysisServer(source).complete(source, offset);
-      log.info('PERF: Computed completions in ${watch.elapsedMilliseconds}ms.');
+      _log.info('PERF: Computed completions in ${watch.elapsedMilliseconds}ms.');
       return response;
     } catch (e, st) {
-      log.severe('Error during _complete', e, st);
+      _log.severe('Error during _complete', e, st);
       await restart();
       rethrow;
     }
@@ -633,7 +412,7 @@ class CommonServer {
     final watch = Stopwatch()..start();
     final response =
         await getCorrectAnalysisServer(source).getFixes(source, offset);
-    log.info('PERF: Computed fixes in ${watch.elapsedMilliseconds}ms.');
+    _log.info('PERF: Computed fixes in ${watch.elapsedMilliseconds}ms.');
     return response;
   }
 
@@ -650,7 +429,7 @@ class CommonServer {
     final watch = Stopwatch()..start();
     final response =
         await getCorrectAnalysisServer(source).getAssists(source, offset);
-    log.info('PERF: Computed assists in ${watch.elapsedMilliseconds}ms.');
+    _log.info('PERF: Computed assists in ${watch.elapsedMilliseconds}ms.');
     return response;
   }
 
@@ -664,7 +443,7 @@ class CommonServer {
 
     final response =
         await getCorrectAnalysisServer(source).format(source, offset);
-    log.info('PERF: Computed format in ${watch.elapsedMilliseconds}ms.');
+    _log.info('PERF: Computed format in ${watch.elapsedMilliseconds}ms.');
     return response;
   }
 
@@ -689,7 +468,7 @@ class CommonServer {
       try {
         await flutterWebManager.initFlutterWeb();
       } catch (e) {
-        log.warning('unable to init package:flutter: $e');
+        _log.warning('unable to init package:flutter: $e');
         return;
       }
     }
